+++ test/cases/exception-execute-asset/webpack.config.js
@@ -1,0 +1,27 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    publicPath: '/',
+  },
+
+  entry: {
+    // test error by execution template function
+    index: 'src/index.pug',
+  },
+
+  plugins: [new PugPlugin({})],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+      },
+    ],
+  },
+};